refactor: move guildcountdowns to same thing as counters
