+++ src/utils/models/Client.ts
@@ -5,11 +5,10 @@ import { doChatReactions } from "../scheduled/clusterjobs/chatreaction";
 import { runModerationChecks } from "../scheduled/clusterjobs/moderationchecks";
 import { updateCache } from "../imghandler";
 import { runLotteryInterval } from "../scheduled/clusterjobs/lottery";
-import { runCountdowns } from "../scheduled/clusterjobs/guildcountdowns";
 import { runChristmas } from "../scheduled/clusterjobs/guildchristmas";
 import { runPremiumChecks } from "../scheduled/clusterjobs/premiumexpire";
 import { runEconomySetup } from "../economy/utils";
-import { updateCounters } from "../guilds/utils";
+import { runCountdowns, updateCounters } from "../guilds/utils";
 
 export class NypsiClient extends Client {
     public cluster: Cluster.Client;
@@ -29,12 +28,12 @@ export class NypsiClient extends Client {
         getWebhooks(this);
         runPremiumChecks();
         updateCounters(this);
+        runCountdowns(this);
 
         if (!this.shard.ids.includes(0)) return;
 
         runLotteryInterval(this);
 
-        runCountdowns(this);
         runChristmas(this);
         runModerationChecks(this);
         doChatReactions(this);
